map is finally back baby

+++ core/src/de/tum/cit/ase/maze/screens/MazeRunnerGame.java
@@ -77,8 +77,8 @@ public class MazeRunnerGame extends Game {
     /**
      * Switches to the game screen.
      */
-    public void goToGame(String mapPath, int score, int time) {
-        gameScreen = new GameScreen(this, mapPath, score, time);
+    public void goToGame(String mapPath,int score,float time) {
+        gameScreen = new GameScreen(this,mapPath,score,time);
         this.setScreen(gameScreen);
     }
     //load game method which uses native file chooser to load in a map from a.properties file
